+++ Stripe/PublicHeaders/STPSourceParams.h
@@ -289,6 +289,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  Creates params for a card source created from Visa Checkout.
+ @see https://stripe.com/docs/visa-checkout
 
  @note Creating an STPSource with these params will give you a
  source with type == STPSourceTypeCard
@@ -297,6 +298,21 @@ NS_ASSUME_NONNULL_BEGIN
  @return An STPSourceParams object populated with the provided values.
  */
 + (STPSourceParams *)visaCheckoutParamsWithCallId:(NSString *)callId;
+
+
+/**
+ Creates params for a card source created from Masterpass.
+ @see https://stripe.com/docs/masterpass
+
+ @note Creating an STPSource with these params will give you a
+ source with type == STPSourceTypeCard
+
+ @param cartId The cartId from a `MCCCheckoutResponse` object.
+ @param transactionId The transactionid from a `MCCCheckoutResponse` object.
+ @return An STPSourceParams object populated with the provided values.
+ */
++ (STPSourceParams *)masterpassParamsWithCartId:(NSString *)cartId
+                                  transactionId:(NSString *)transactionId;
 
 @end